위도경도 좌표 float 정밀도로 인한 double로 변경.

+++ src/main/java/com/robi/foodiy/data/dao/RecordsDao.java
@@ -14,8 +14,8 @@ public class RecordsDao {
     private Date whenDate;          // 일자(yyyy-MM-dd)
     private Time whenTime;          // 시간(HH:mm:ss)
     private String wherePlace;      // 장소명
-    private Float whereLati;        // 위도
-    private Float whereLongi;       // 경도
+    private Double whereLati;        // 위도
+    private Double whereLongi;       // 경도
     private String whoWith;         // 같이있던 사람
     private String picUrls;         // 사진URL들
 }
